- Removed `MappedSchema` - Moved `CordaPersistenceException` and `UUIDConverter` into application module - Removed persistence module - Fixed up obsolete ledger module to support these changes.
