+++ application/src/main/kotlin/net/corda/v5/application/persistence/PersistenceService.kt
@@ -2,7 +2,6 @@ package net.corda.v5.application.persistence
 
 import net.corda.v5.base.annotations.DoNotImplement
 import net.corda.v5.base.annotations.Suspendable
-import net.corda.v5.persistence.CordaPersistenceException
 
 /**
  * Persistence Service API providing functionality to interact with an entityManager and execute pre-defined named queries.
