+++ application/src/main/kotlin/net/corda/v5/application/persistence/jpa/UUIDConverter.kt
@@ -1,10 +1,10 @@
-package net.corda.v5.persistence
+package net.corda.v5.application.persistence.jpa
 
 import java.util.UUID
 import javax.persistence.AttributeConverter
 
-/** Used to convert to and from [UUID]s when defining [MappedSchema] entity columns. */
-object UUIDConverter : AttributeConverter<UUID, String> {
+/** [AttributeConverter] used to convert JPA Entity String types to DB [UUID] types and vice versa. */
+class UUIDConverter : AttributeConverter<UUID, String> {
     override fun convertToDatabaseColumn(uuid: UUID?) = uuid?.toString()
     override fun convertToEntityAttribute(dbData: String?) = if (dbData != null) UUID.fromString(dbData) else null
 }
