+++ cordapp-configuration/src/main/resources/net/corda/cordapp/cordapp-configuration.properties
@@ -1,6 +1,5 @@
 Corda-Contract-Classes=IMPLEMENTS;net.corda.v5.ledger.contracts.Contract
 Corda-Flow-Classes=IMPLEMENTS;net.corda.v5.application.flows.Flow
-Corda-MappedSchema-Classes=EXTENDS;net.corda.v5.persistence.MappedSchema
 Corda-SerializationWhitelist-Classes=IMPLEMENTS;net.corda.v5.serialization.SerializationWhitelist
 Corda-SerializationCustomSerializer-Classes=IMPLEMENTS;net.corda.v5.serialization.SerializationCustomSerializer
 Corda-CheckpointCustomSerializer-Classes=IMPLEMENTS;net.corda.v5.serialization.CheckpointCustomSerializer
