--- persistence/build.gradle
@@ -1,17 +1,0 @@
-plugins {
-    id 'corda-api.common-library'
-    id 'corda.common-publishing'
-    id 'org.jetbrains.kotlin.jvm'
-}
-
-description 'Corda Persistence'
-
-dependencies {
-    compileOnly 'org.osgi:osgi.annotation'
-
-    api 'javax.persistence:javax.persistence-api'
-    implementation 'net.corda.kotlin:kotlin-stdlib-jdk8-osgi'
-
-    implementation platform(project(':corda-api'))
-    implementation project(":base")
-}
