--- persistence/src/main/java/net/corda/v5/persistence/package-info.java
@@ -1,5 +1,0 @@
-@Export
-package net.corda.v5.persistence;
-
-import org.osgi.annotation.bundle.Export;
-
